moves bernoulli-specific functions to their own file
