+++ CMakeLists.txt
@@ -59,6 +59,7 @@ pybind11_add_module(_pdaggerq
         pdaggerq/pq_tensor.cc
         pdaggerq/pq_string.cc
         pdaggerq/pq_utils.cc
+        pdaggerq/pq_bernoulli.cc
         pdaggerq/pq_serialize.cc
         pdaggerq/pq_swap_operators.cc
         pdaggerq/pq_add_spin_labels.cc
